// src/components/enums/Filter.ts
export enum Filter {
  Female = "TIK",
  Male = "HANE",
  Small = "LITEN",
  Medium = "MELLAN",
  Big = "STOR",
  Puppy = "VALP",
  Adult = "VUXEN",
  Senior = "SENIOR",
  Clear = "RENSA",
}

// export enum Filter {
//   Female = "TIK",
//   Male = "HANE",
//   Small = "LITEN < 8KG",
//   Medium = "MELLAN 9-25KG",
//   Big = "STOR > 25KG",
//   Puppy = "VALP < 1 ÅR",
//   Adult = "VUXEN 1-5 ÅR",
//   Senior = "SENIOR > 6 ÅR",
//   Clear = "RENSA",
// }

// export enum Filter {
//   Female = "Tik",
//   Male = "Hane",
//   Small = "Liten < 8kg",
//   Medium = "Mellan 9-25kg",
//   Big = "Stor > 25kg",
//   Puppy = "Valp < 1 år",
//   Adult = "Vuxen 1-5 år",
//   Senior = "Senior > 6 år",
//   Clear = "Rensa",
// }